Criada classe User e seu respectivo DAO - Long mudado para Int nos ID

+++ app/src/main/java/me/emersondantas/coursescounter/model/dao/CourseDAO.java
@@ -12,7 +12,7 @@ import me.emersondantas.coursescounter.model.bean.Course;
  * @author Emerson Dantas
  */
 public class CourseDAO extends DataAccessObject{
-    private long lastId;
+    private int lastId;
     private static CourseDAO instance;
 
     private CourseDAO() {
